archivo configuraciones y anda movimiento

+++ TurnEvent.h
@@ -8,12 +8,12 @@ enum TurningSense {ANTICLOCKWISE, CLOCKWISE};
 
 class TurnEvent: public AbstractEvent {
     TurningSense sense;
-    int degrees;
+    double degrees;
 public:
-    explicit TurnEvent(TurningSense sense, int degrees): sense(sense), degrees(degrees) {}
+    explicit TurnEvent(TurningSense sense, double degrees): sense(sense), degrees(degrees) {}
     void runHandler(GameStage& gameStage) override;
     TurningSense getSense();
-    int getDegrees();
+    double getDegrees();
 };
 
 #endif
